objtool: Skip files with existing orc section silently
When rebuilding an LTO tree the build system often decides
to rerun objtool orc on an existing .lto.o file again.
objtool then errors out if the orc section already exists
and fails the build.

Just silently skip regenerating orc if it already exists.

Maybe that would be better fixed in the build system,
but I'm not fully sure how.

Signed-off-by: Andi Kleen <andi@firstfloor.org>